requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha>=2.0.0a1
boto3>=1.28.0